+++ app/Jobs/Git/SshAndGitPull.php
@@ -252,7 +252,7 @@ class SshAndGitPull implements ShouldQueue
         if ( $success )
         {
             // Dispatch event for git pull success
-            event(new GitPullSuccess( $this->repository ) );
+            event(new GitPullSuccess( $this->repository, $this->site, $this->deployment_type ) );
             // Save the db
             $this->status = RepoStatus::SUCCESS->value;
             $this->status_text = 'Git pull success.';